requirements.txt
absl-py==0.9.0
asgiref==3.2.5
astor==0.8.1
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
httpie==2.1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
mkl-service==2.3.0
Pillow==7.1.0
protobuf==3.11.4
psycopg2==2.8.4
Pygments==2.6.1
pylint==2.4.4
pyreadline==2.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1